Tests/LibWeb/Text/expected/collapse-selection-inside-shadow-root.txt

null
null
button1
button1
button2
button2
